Extract to util/auth.ts, add maxRefetchAttempts = 5
